Refactor routes and controllers to the middleware folder
